\name{datap}

\docType{data}

\alias{datap}

\title{Test Panel Data Set: Cultural Values, GDP and Democracy from 1981-2012 in 74 countries}

\description{
\code{datap} is a panel data frame, that was produced with \code{plm}{pdata.frame}. It is used as testdataset for the R package \code{bdynsys}. It contains 7 variables. The first is the panel_id variable with year and country specification. The six other variables are: EmanzV (emancipative values, based on WorldValueSurvey), SeculTradV (secular_vs_traditional values, based on WorldValueSurvey), SurvSelfV (survival_vs_selfexpression values, based on WorldValueSurvez), DemocrH (human-rights weighted democracy index, based on Freedom House data), GDP (GDP per capita, based on World Bank data) and logGDP (logged and rescaled (0-1) GDP per capita). }

\references{
       Wezel, C. (2013) \emph{Freedom Rising}.
       Cambridge: Cambridge University Press.
     }

\format{panel data frame}

\keyword{datasets}