\name{datap}

\docType{data}

\alias{datap}

\title{Test Panel Data Set: Cultural Values, GDP and Democracy from 1981-2012 in 74 countries}

\description{
\code{datap} is a panel data frame, that was produced with \code{plm}{pdata.frame}. It is used as testdataset for the R package \code{bdynsys}. It contains 7 variables. The first is the panel_id variable with year and country specification. The six other variables are: EmanzV (emancipative values, based on WorldValueSurvey), SeculTradV (secular_vs_traditional values, based on WorldValueSurvey), SurvSelfV (survival_vs_selfexpression values, based on WorldValueSurvez), DemocrH (human-rights weighted democracy index, based on Freedom House data), GDP (GDP per capita, based on World Bank data) and logGDP (logged and rescaled (0-1) GDP per capita). The countries IDs in the data are as follows:
1: Albania, 2: Argentina, 3: Armenia, 4: Australia, 5: Austria, 6: Azerbaijan, 7: Bangladesh, 8: Belarus, 9: Belgium, 10: Bosnia and Herzigowina, 11: Brazil, 12: Bulgaria, 13: Canada, 14: Chile, 15: China, 16: Colombia, 17: Croatia, 18: Cyprus, 19: Czech Republik, 20: Denmark, 21: Egypt, 22: Estonia, 23: Finland, 24: France, 25: Georgia, 26: Germany, 27: Greece, 28: Hungary, 29: Iceland, 30: India, 31: Indonesia, 32: Iran, 33: Iraq, 34: Ireland, 35: Israel, 36: Italy, 37: Japan, 38: Jordan, 39: South Korea, 40: Latvia, 41: Lithunia, 42: Luxembourg, 43: Macedonia, 44: Malta, 45: Mexico, 46: Moldova, 47: Montenegro, 48: Morocco, 49: Netherlands, 50: New Zealand, 51: Nigeria, 52: Norway, 53: Pakistan, 54: Peru, 55: Philippines, 56: Poland, 57: Portugal, 58: Puerto Rico, 59: Romania, 60: Russian Federation, 61: Serbia, 62: Slovak Republic, 63: Slovenia, 64: South Africa, 65: Spain, 66: Sweden, 67: Switzerland, 68: Turkey, 69: Ukraine, 70: United Kingdom, 71: USA, 72: Uruguay, 73: Venezuela, 74: Vietnam}

\references{
       Wezel, C. (2013) \emph{Freedom Rising}.
       Cambridge: Cambridge University Press.
     }

\format{panel data frame}

\keyword{datasets}